refactor(server): consolidate the two server entrypoints into one
server.ts is now the single entry for both backends, branching on DATABASE_URL:
postgres:// connects to an external Postgres; a path/file:// spawns a local
embedded Postgres. The embedded boot lives in the new embedded-runtime.ts and
is loaded ONLY via `await import(...)` in the embedded branch, so the
external/prod path never resolves or loads the embedded-postgres binary even
though it sits in node_modules.

- Deleted start-local.ts; build one bundle (server.bundle.mjs) instead of two.
- dev.ts drops bundle-switching (one bundle, self-selects on DATABASE_URL) and
  maps LOBU_DATA_DIR → an embedded DATABASE_URL, so the Mac app (which spawns
  `lobu run` with LOBU_DATA_DIR) keeps working with no Swift change.
- dev:local + e2e-lobu-apply.sh point at server.ts.

Validated both modes from the one bundle: embedded (file://) spawns PG, runs
migrations, all 4 extensions live; external (postgres://) connects, passes the
schema-version check, and never triggers the lazy embedded-postgres import
(0 spawn lines). typecheck 0, gateway tests 24/24.
