+++ scripts/dev-native.sh
@@ -130,7 +130,7 @@ mkdir -p "$LOBU_WORKSPACE_ROOT"
 # --- Run -------------------------------------------------------------------
 
 if [ -z "${DATABASE_URL:-}" ]; then
-  # No external Postgres → boot the embedded Postgres backend (src/start-local.ts).
+  # No external Postgres → boot the embedded Postgres backend (src/server.ts).
   # First run mints a web login (dev@lobu.local / lobudev123, org "dev") and a
   # bootstrap PAT under LOBU_DATA_DIR. Vite HMR still runs in-process.
   export LOBU_DATA_DIR="${LOBU_DATA_DIR:-$REPO_ROOT/.lobu-dev}"
